Add user horrortroll in userspace (#19769)
