+++ keyboards/gopolar/gg86/keymaps/bongocat/readme.md
@@ -1,3 +1,1 @@
 # Keymap with VIA and Bongo Cat on OLED
-
-Keymap is default 87 qwerty, with TKL 7u-spacebar layout
