+++ keyboards/gopolar/gg86/keymaps/horrortroll/keymap.c
@@ -1,4 +1,4 @@
-/* Copyright 2021 HorrorTroll <https://github.com/HorrorTroll>
+/* Copyright 2023 HorrorTroll <https://github.com/HorrorTroll>
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -16,7 +16,8 @@
 
 #include QMK_KEYBOARD_H
 
-#include "keymap_stuff.h"
+#include "horrortroll.h"
+#include "oled/oled_stuff.h"
 
 const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
