--- keyboards/gopolar/gg86/keymaps/horrortroll/led/cool_diagonal.c
@@ -1,22 +1,0 @@
-/* Copyright 2021 HorrorTroll <https://github.com/HorrorTroll>
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 2 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-static HSV COOL_DIAGONAL_math(HSV hsv, uint8_t i, uint8_t time) {
-    hsv.h = (g_led_config.point[i].x / 4) - g_led_config.point[i].y - time;
-    return hsv;
-}
-
-bool COOL_DIAGONAL(effect_params_t* params) { return effect_runner_i(params, &COOL_DIAGONAL_math); }
