--- keyboards/gopolar/gg86/keymaps/horrortroll/readme.md
@@ -1,12 +1,0 @@
-# My personal keymap with VIA and custom LED / OLED
-
-Keymap is default 87 qwerty, with TKL 7u-spacebar layout
-
-It have new LED effect:
-
-- Custom gradient (ported from SirTimmyTimbit code [https://github.com/SirTimmyTimbit/customizable-gradient-effect-for-drop-alt])
-- Cool diagonal (ported from pleasuretek code [https://github.com/pleasuretek/qmk_firmware])
-- Flower Blooming
-- Rainbow reactive simple
-- Knight Rider (ported from jumper149 code [https://github.com/jumper149/qmk_firmware/blob/jumper149/keyboards/dztech/dz65rgb/keymaps/jumper149/])
-- Random breath rainbow (based from daed code [https://github.com/daed/qmk_firmware/blob/master/keyboards/massdrop/alt/keymaps/daed] and modify by me)
