--- keyboards/gopolar/gg86/keymaps/horrortroll/rules.mk
@@ -1,3 +1,0 @@
-VIA_ENABLE = yes
-
-RGB_MATRIX_CUSTOM_USER = yes
