+++ users/horrortroll/led/cool_diagonal.c
